+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- [#363](https://github.com/loop-payments/prisma-lint/issues/363) Require enum fields to use snake case in `field-name-mapping-snake-case`.
+- [#363](https://github.com/loop-payments/prisma-lint/issues/363) Require enum and custom type fields to use snake case in `field-name-mapping-snake-case`.
 
 ## 0.3.0 (2024-04-03)